Fixtures name autogeneration adjustments
